# [skrifa] path conversion perf improvements