--- a/fauntlet/src/compare_glyphs.rs
+++ b/fauntlet/src/compare_glyphs.rs
@@ -40,36 +40,35 @@ pub fn compare_glyphs(
 
     for gid in 0..glyph_count {
         let gid = GlyphId::from(gid);
-        // Restore this when <https://github.com/googlefonts/fontations/issues/790>
-        // is completed.
-        //
-        // let ft_advance = ft_instance.advance(gid);
-        // let skrifa_advance = skrifa_instance.advance(gid);
-        // if ft_advance != skrifa_advance {
-        //     writeln!(
-        //         std::io::stderr(),
-        //         "[{path:?}#{} ppem: {} coords: {:?}] glyph id {} advance doesn't match:\nFreeType: {ft_advance:?}\nSkrifa:   {skrifa_advance:?}",
-        //         options.index,
-        //         options.ppem,
-        //         options.coords,
-        //         gid.to_u16(),
-        //     )
-        //     .unwrap();
-        //     if exit_on_fail {
-        //         std::process::exit(1);
-        //     }
-        // }
         ft_outline.clear();
-        ft_instance
+        let ft_advance = ft_instance
             .outline(gid, &mut RegularizingPen::new(&mut ft_outline, is_scaled))
             .unwrap();
         skrifa_outline.clear();
-        skrifa_instance
+        let maybe_skrifa_advance = skrifa_instance
             .outline(
                 gid,
                 &mut RegularizingPen::new(&mut skrifa_outline, is_scaled),
             )
             .unwrap();
+        // Compare against adjusted metrics when skrifa returns them (currently
+        // only for TrueType glyphs)
+        if let Some(skrifa_advance) = maybe_skrifa_advance {
+            if ft_advance != skrifa_advance {
+                writeln!(
+                std::io::stderr(),
+                "[{path:?}#{} ppem: {} coords: {:?}] glyph id {} advance doesn't match:\nFreeType: {ft_advance:?}\nSkrifa:   {skrifa_advance:?}",
+                options.index,
+                options.ppem,
+                options.coords,
+                gid.to_u32(),
+            )
+            .unwrap();
+                if exit_on_fail {
+                    std::process::exit(1);
+                }
+            }
+        }
         if ft_outline != skrifa_outline {
             ok = false;
             fn outline_to_string(outline: &[PathElement]) -> String {