Reworks the TrueType->path conversion code to capture states explicitly to reduce branches along with some inlining and other code motion.

This gives a roughly 10% improvement based on measured fonts.

Example for Noto Sans Regular (all glyphs 100x):
| FreeType | skrifa (current) | skrifa (this PR) | skrifa (this PR + LTO) |
|-----------|-------------|--------------|--------------------|
| 308ms     | 337ms      |  295ms        | 213ms                  |

The significant difference when enabling LTO suggests that we might be missing some inlining opportunities across the `read-fonts`/`font-types` boundaries. Note that current skrifa + LTO is roughly the same as this PR without it.