@dfrg 2024-10-03
> I think the target branch is set wrong on this PR, shouldn't it be main?

I chained this locally because I wanted to capture the advance with changes in my performance measurements. Should be based on main now that the other PR is merged.